chore: move custom navigator to rai.extensions
